Cache the tags extracted from connection strings
